# 🔐 Production-Ready OAuth UX & Service Dependency Management